+++ src/infrastructure/repositories/adminRepository.ts
@@ -1,4 +1,4 @@
-import { IAdminRepo } from "../../interfaces/repos/adminRepo";
+import { IAdminRepo } from "../../useCases/repos/adminRepo";
 import { IAdmin } from "../../interfaces/schema/adminSchema";
 import { adminModel } from "../db/adminModel";
