+++ src/useCases/repos/adminRepo.ts
@@ -1,4 +1,4 @@
-import { IAdmin } from "../schema/adminSchema"
+import { IAdmin } from "../../interfaces/schema/adminSchema"
 
 export interface IAdminRepo {
     findAdmin(): Promise<IAdmin | null>
