+++ src/useCases/repos/couponRepo.ts
@@ -1,4 +1,4 @@
-import { ICouponReqs, ICouponRes } from "../schema/couponSchema";
+import { ICouponReqs, ICouponRes } from "../../interfaces/schema/couponSchema";
 
 export interface ICouponRepo {
     addCoupon (coupon: ICouponReqs): Promise<ICouponRes>
