+++ src/useCases/repos/screenSeatRepo.ts
