+++ src/useCases/repos/showRepo.ts
@@ -1,4 +1,4 @@
-import { IShow, IShowToSave } from "../schema/showSchema";
+import { IShow, IShowToSave } from "../../interfaces/schema/showSchema";
 
 export interface IShowRepo {
     saveShow (showTosave: IShowToSave): Promise<IShow>
