+++ src/useCases/repos/tempTicketRepo.ts
@@ -1,4 +1,4 @@
-import { ITempTicketRes, ITempTicketReqs } from "../schema/ticketSchema";
+import { ITempTicketRes, ITempTicketReqs } from "../../interfaces/schema/ticketSchema";
 
 export interface ITempTicketRepo {
     saveTicketDataTemporarily (ticketData: ITempTicketReqs): Promise<ITempTicketRes>
